requirements.txt
pymilvus
sentence_transformers
tk
torch
fairscale
fire
sentencepiece
openai